feat: add Slidev HTML to PPTX conversion prototype
Implements an end-to-end pipeline demonstrating the conversion of
Slidev-style HTML presentations to editable PowerPoint format.

Pipeline:
- Input PPTX created with PptxGenJS (5 slides)
- HTML representation with Slidev-style CSS
- dom-to-pptx simulation using PptxGenJS
- Output PPTX with preserved styling

Features:
- 5 sample slides: title, bullet points, two-column layout, code block, thank you
- Browserless integration for HTML screenshot capture at 1920x1080
- Visual comparison report with side-by-side analysis
- Python script for PPTX-to-image conversion attempt

Also fixes browserless screenshot quality parameter (Int -> Float) to
match current BrowserQL API.

+++ browserless/src/browserless-client.ts
@@ -233,7 +233,7 @@ export async function screenshot(options: ScreenshotOptions = {}): Promise<strin
   } = options;
 
   const query = `
-    mutation Screenshot($type: ScreenshotType!, $quality: Int!, $fullPage: Boolean!, $optimizeForSpeed: Boolean!) {
+    mutation Screenshot($type: ScreenshotType!, $quality: Float!, $fullPage: Boolean!, $optimizeForSpeed: Boolean!) {
       screenshot(type: $type, quality: $quality, fullPage: $fullPage, optimizeForSpeed: $optimizeForSpeed) {
         base64
       }
